Moved ObjectStore tests to datatypes_test.py
